Fix: 1.修复微信支付在4s下崩溃的问题（SCWeiXinPayOder模型里timestamp属性申明由UInt32变为NSInteger，避免在32位机器上解析引起崩溃）

+++ MaintenanceCar/Models/Merchant/GroupProduct/SCWeiXinPayOder.h
@@ -18,6 +18,6 @@
 @property (nonatomic, copy)   NSString <Optional>*prepayid;
 @property (nonatomic, copy)   NSString <Optional>*partnerid;
 @property (nonatomic, copy)   NSString <Optional>*out_trade_no;
-@property (nonatomic, assign) UInt32             timestamp;
+@property (nonatomic, assign) NSInteger           timestamp;
 
 @end
